+++ Libraries/Opc.Ua.Security.Certificates/Opc.Ua.Security.Certificates.csproj
@@ -31,9 +31,6 @@
     <PackageReference Include="Portable.BouncyCastle" Version="1.8.8" />
   </ItemGroup>
 
-  <ItemGroup Condition="'$(TargetFramework)' == 'netstandard2.1'">
-  </ItemGroup>
-
   <Target Name="GetPackagingOutputs" />
 
 </Project>
