+++ Libraries/Opc.Ua.Security.Certificates/Org.BouncyCastle/CertificateFactoryRandomGenerator.cs
@@ -1,0 +1,75 @@
+/* Copyright (c) 1996-2020 The OPC Foundation. All rights reserved.
+   The source code in this file is covered under a dual-license scenario:
+     - RCL: for OPC Foundation members in good-standing
+     - GPL V2: everybody else
+   RCL license terms accompanied with this source code. See http://opcfoundation.org/License/RCL/1.00/
+   GNU General Public License as published by the Free Software Foundation;
+   version 2 of the License are accompanied with this source code. See http://opcfoundation.org/License/GPLv2
+   This source code is distributed in the hope that it will be useful,
+   but WITHOUT ANY WARRANTY; without even the implied warranty of
+   MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.
+*/
+
+#if !NETSTANDARD2_1
+using System;
+using System.Security.Cryptography;
+using Org.BouncyCastle.Crypto.Prng;
+
+namespace Opc.Ua.Security.Certificates.BouncyCastle
+{
+    /// <summary>
+    /// Secure .Net Core Random Number generator wrapper for Bounce Castle.
+    /// Creates an instance of RNGCryptoServiceProvider or an OpenSSL based version on other OS.
+    /// </summary>
+    public class CertificateFactoryRandomGenerator : IRandomGenerator, IDisposable
+    {
+        RandomNumberGenerator m_prg;
+
+        /// <summary>
+        /// Creates an instance of a crypthographic secure random number generator.
+        /// </summary>
+        public CertificateFactoryRandomGenerator()
+        {
+            m_prg = RandomNumberGenerator.Create();
+        }
+
+        /// <summary>
+        /// Dispose the random number generator.
+        /// </summary>
+        public void Dispose()
+        {
+            m_prg.Dispose();
+        }
+
+        /// <summary>Add more seed material to the generator. Not needed here.</summary>
+        public void AddSeedMaterial(byte[] seed) { }
+
+        /// <summary>Add more seed material to the generator. Not needed here.</summary>
+        public void AddSeedMaterial(long seed) { }
+
+        /// <summary>
+        /// Fills an array of bytes with a cryptographically strong
+        /// random sequence of values.
+        /// </summary>
+        /// <param name="bytes">Array to be filled.</param>
+        public void NextBytes(byte[] bytes)
+        {
+            m_prg.GetBytes(bytes);
+        }
+
+        /// <summary>
+        /// Fills an array of bytes with a cryptographically strong
+        /// random sequence of values.
+        /// </summary>
+        /// <param name="bytes">Array to receive bytes.</param>
+        /// <param name="start">Index to start filling at.</param>
+        /// <param name="len">Length of segment to fill.</param>
+        public void NextBytes(byte[] bytes, int start, int len)
+        {
+            byte[] temp = new byte[len];
+            m_prg.GetBytes(temp);
+            Array.Copy(temp, 0, bytes, start, len);
+        }
+    }
+}
+#endif
