+++ Libraries/Opc.Ua.Security.Certificates/Org.BouncyCastle/X509Utils.cs
@@ -19,7 +19,6 @@ using System.Security.Cryptography.X509Certificates;
 using Org.BouncyCastle.Asn1.X509;
 using Org.BouncyCastle.Crypto;
 using Org.BouncyCastle.Crypto.Parameters;
-using Org.BouncyCastle.Crypto.Prng;
 using Org.BouncyCastle.Math;
 using Org.BouncyCastle.OpenSsl;
 using Org.BouncyCastle.Pkcs;
@@ -27,97 +26,6 @@ using Org.BouncyCastle.Security;
 
 namespace Opc.Ua.Security.Certificates.BouncyCastle
 {
-    /// <summary>
-    /// Secure .Net Core Random Number generator wrapper for Bounce Castle.
-    /// Creates an instance of RNGCryptoServiceProvider or an OpenSSL based version on other OS.
-    /// </summary>
-    public class CertificateFactoryRandomGenerator : IRandomGenerator, IDisposable
-    {
-        RandomNumberGenerator m_prg;
-
-        /// <summary>
-        /// Creates an instance of a crypthographic secure random number generator.
-        /// </summary>
-        public CertificateFactoryRandomGenerator()
-        {
-            m_prg = RandomNumberGenerator.Create();
-        }
-
-        /// <summary>
-        /// Dispose the random number generator.
-        /// </summary>
-        public void Dispose()
-        {
-            m_prg.Dispose();
-        }
-
-        /// <summary>Add more seed material to the generator. Not needed here.</summary>
-        public void AddSeedMaterial(byte[] seed) { }
-
-        /// <summary>Add more seed material to the generator. Not needed here.</summary>
-        public void AddSeedMaterial(long seed) { }
-
-        /// <summary>
-        /// Fills an array of bytes with a cryptographically strong
-        /// random sequence of values.
-        /// </summary>
-        /// <param name="bytes">Array to be filled.</param>
-        public void NextBytes(byte[] bytes)
-        {
-            m_prg.GetBytes(bytes);
-        }
-
-        /// <summary>
-        /// Fills an array of bytes with a cryptographically strong
-        /// random sequence of values.
-        /// </summary>
-        /// <param name="bytes">Array to receive bytes.</param>
-        /// <param name="start">Index to start filling at.</param>
-        /// <param name="len">Length of segment to fill.</param>
-        public void NextBytes(byte[] bytes, int start, int len)
-        {
-            byte[] temp = new byte[len];
-            m_prg.GetBytes(temp);
-            Array.Copy(temp, 0, bytes, start, len);
-        }
-    }
-
-    /// <summary>
-    /// A converter class to create a X509Name object 
-    /// from a X509Certificate subject.
-    /// </summary>
-    /// <remarks>
-    /// Handles subtle differences in the string representation
-    /// of the .NET and the Bouncy Castle implementation.
-    /// </remarks>
-    public class CertificateFactoryX509Name : X509Name
-    {
-        /// <summary>
-        /// Create the X509Name from a distinguished name.
-        /// </summary>
-        /// <param name="distinguishedName">The distinguished name.</param>
-        public CertificateFactoryX509Name(string distinguishedName) :
-            base(true, ConvertToX509Name(distinguishedName))
-        {
-        }
-
-        /// <summary>
-        /// Create the X509Name from a distinguished name.
-        /// </summary>
-        /// <param name="reverse">Reverse the order of the names.</param>
-        /// <param name="distinguishedName">The distinguished name.</param>
-        public CertificateFactoryX509Name(bool reverse, string distinguishedName) :
-            base(reverse, ConvertToX509Name(distinguishedName))
-        {
-        }
-
-        private static string ConvertToX509Name(string distinguishedName)
-        {
-            // convert from X509Certificate to bouncy castle DN entries
-            return distinguishedName.Replace("S=", "ST=");
-        }
-    }
-
     /// <summary>
     /// Wrapper for a password string.
     /// </summary>
